Avoid trying to compile Linux-only features on macOS
This is a blunt instrument, we merely conditionally compile out things
that macOS doesn't support.  But it's enough to be able to build and run
on macOS.

I've assumed a lax support policy, please feel free to commit to a
stricter policy if you so desire :).

I also fixed up tests that use DNS to not hit a real resolver - macOS's
resolver doesn't seem to randomise results on the timescale needed for
tests, and it's cleaner to not rely on real nameservers if we don't have
to.