Fix minimum MAC size this fixes a deserialization error
Before the last extension was always ignored if it was not at least 32
bytes without padding.

Removed a test that did check for the wrong behavior
This test checked that a too short field was not actually serialized at
all, which is wrong. Making the encrypted field this short is also not
possible as the nonce alone is 16bytes + 4bytes of ef header + 4 bytes
NTS header...